refactor changes by kikkia
